requirements.txt
Flask
flask-cors
python-dotenv
opencv-python-headless
numpy
easyocr
ultralytics
requests
urllib3
